feat: Arcade Scanner improvements and tests
