+++ arcade_scanner/server/api_handler.py
@@ -292,11 +292,12 @@ class FinderHandler(http.server.SimpleHTTPRequestHandler):
 
     def do_GET(self):
         try:
-            from .routes import queue, settings, duplicates, tags
+            from .routes import queue, settings, duplicates, tags, files
             if queue.handle_get(self): return
             if settings.handle_get(self): return
             if duplicates.handle_get(self): return
             if tags.handle_get(self): return
+            if files.handle_get(self): return
         except Exception as e:
             print(f"Module route error GET: {e}")
         try:
@@ -530,548 +531,17 @@ class FinderHandler(http.server.SimpleHTTPRequestHandler):
                 else:
                      self.send_error(404, "User not found")
                 return
-            
-            elif self.path.startswith("/reveal?"):
-                try:
-                    params = parse_qs(urlparse(self.path).query)
-                    file_path = params.get("path", [None])[0]
-                    if not file_path:
-                         self.send_error(400, "Missing path parameter")
-                         return
-                         
-                    print(f"🔍 Reveal requested for: {file_path}")
 
-                    # Check if file is in a hidden folder (starts with .)
-                    abs_path = os.path.abspath(file_path)
-                    is_hidden = any(part.startswith('.') for part in Path(abs_path).parts if part != '/')
-
-                    if is_hidden:
-                        # File exists in hidden folder - return helpful response instead of blocking
-                        print(f"📁 File in hidden folder: {abs_path}")
-                        self.send_response(200)
-                        self.send_header('Content-Type', 'application/json')
-                        self.end_headers()
-                        response = json.dumps({
-                            "status": "hidden_folder",
-                            "path": abs_path,
-                            "message": "This file is located in a hidden system folder"
-                        })
-                        self.wfile.write(response.encode())
-                        return
-
-                    # Security Fix C-3: Validate path is within allowed directories
-                    if not is_path_allowed(file_path):
-                        print(f"🚨 Unauthorized reveal attempt blocked: {file_path}")
-                        self.send_error(403, "Forbidden - Path not in scan directories")
-                        return
-
-                    if not os.path.exists(file_path):
-                        print(f"❌ Error: File does not exist: {file_path}")
-                        self.send_error(404, "File not found")
-                        return
-
-                    if IS_WIN:
-                        subprocess.run(["explorer", "/select,", os.path.normpath(file_path)])
-                    elif sys.platform == "darwin":
-                        print(f"🚀 Running: open -R '{file_path}'")
-                        result = subprocess.run(["open", "-R", file_path], capture_output=True, text=True)
-                        if result.returncode != 0:
-                            print(f"❌ Error revealing file: {result.stderr}")
-                        else:
-                            print("✅ Reveal command successful")
-                    else:
-                        # Linux / Other: Open parent directory since standard reveal is non-standard
-                        parent_dir = os.path.dirname(file_path)
-                        print(f"🚀 Running: xdg-open '{parent_dir}'")
-                        subprocess.run(["xdg-open", parent_dir])
-
-                    self.send_response(204)
-                    self.end_headers()
-                except SecurityError as e:
-                    print(f"🚨 Security violation: {e}")
-                    self.send_error(403, "Forbidden")
-                except Exception as e:
-                    print(f"❌ Critical error in reveal endpoint: {e}")
-                    self.send_error(500, str(e))
-            elif self.path.startswith("/api/mark_optimized?"):
-                params = parse_qs(urlparse(self.path).query)
-                path = params.get("path", [None])[0]
-                if path:
-                    abs_path = os.path.abspath(path)
-                    
-                    # Use new DB layer
-                    entry = db.get(abs_path)
-                    if entry:
-                        entry.status = "OK"
-                    else:
-                        # Create new if missing
-                        size_mb = 0
-                        try:
-                            if os.path.exists(abs_path):
-                                size_mb = os.path.getsize(abs_path) / (1024 * 1024)
-                        except OSError as e:
-                            print(f"⚠️ Could not stat file {abs_path}: {e}")
-                        from arcade_scanner.models.video_entry import VideoEntry
-                        entry = VideoEntry(
-                            FilePath=abs_path,
-                            Size_MB=size_mb,
-                            Status="OK"
-                        )
-                    db.upsert(entry)
-                    db.save()
-                    _media_cache.invalidate()  # Entry geändert → Cache leeren
-                    
-                    # Regenerate HTML report so refresh works
-                    try:
-                        current_port = self.server.server_address[1]
-                        report_debouncer.schedule(current_port)
-                        print(f"✅ Marked as optimized and report update scheduled: {os.path.basename(abs_path)}")
-                    except Exception as e:
-                        print(f"⚠️ Cache updated but report scheduling failed: {e}")
-
-                self.send_response(204)
-                self.end_headers()
-
-            elif self.path.startswith("/compress?"):
-                user_name = self.get_current_user()
-                if not user_name:
-                    self.send_error(401, "Unauthorized")
-                    return
-
-                try:
-                    params = parse_qs(urlparse(self.path).query)
-                    file_path = params.get("path", [None])[0]
-                    
-                    if not file_path:
-                         print("❌ No path provided for compression")
-                         self.send_error(400, "Missing path parameter")
-                         return
-                    
-                    # Security Fix C-1: Sanitize and validate path
-                    try:
-                        file_path = sanitize_path(file_path)
-                    except (SecurityError, ValueError) as e:
-                        print(f"🚨 Security violation in compress: {e}")
-                        self.send_error(403, "Forbidden - Invalid path")
-                        return
-
-                    audio_mode = params.get("audio", ["enhanced"])[0]
-                    video_mode = params.get("video", ["compress"])[0]
-                    q_val = params.get("q", [None])[0]
-                    ss = params.get("ss", [None])[0]  # Trim start time
-                    to = params.get("to", [None])[0]  # Trim end time
-                    
-                    # Validate audio_mode (whitelist)
-                    if audio_mode not in ["enhanced", "standard"]:
-                        print(f"🚨 Invalid audio mode: {audio_mode}")
-                        self.send_error(400, "Invalid audio mode")
-                        return
-                        
-                    # Validate video_mode
-                    if video_mode not in ["compress", "copy"]:
-                        print(f"🚨 Invalid video mode: {video_mode}")
-                        self.send_error(400, "Invalid video mode")
-                        return
-                    
-                    # Get current running port
-                    current_port = self.server.server_address[1]
-                    print(f"🔌 Current Server Port: {current_port}")
-                    print(f"⚡ Optimize: {file_path} | Video: {video_mode} | Audio: {audio_mode} | Q: {q_val} | Trim: {ss}-{to}")
-                    
-                    # Build command as list (NEVER use shell=True!)
-                    cmd_parts = [sys.executable, config.optimizer_path, file_path,
-                                 "--port", str(current_port),
-                                 "--audio-mode", audio_mode,
-                                 "--video-mode", video_mode]
-                    
-                    if ss: 
-                        cmd_parts.extend(["--ss", ss])
-                    if to:
-                        cmd_parts.extend(["--to", to])
-                    if q_val:
-                         cmd_parts.extend(["--q", q_val])
-                    
-                    if IS_WIN:
-                        # Windows: Launch in new console WITHOUT shell=True
-                        print(f"🚀 Launching Optimizer (Win): {' '.join(cmd_parts)}")
-                        subprocess.Popen(
-                            cmd_parts,
-                            creationflags=subprocess.CREATE_NEW_CONSOLE
-                        )
-                    else:
-                        # macOS: Use shlex.quote() for safe AppleScript string building
-                        safe_cmd = ' '.join(shlex.quote(str(p)) for p in cmd_parts)
-                        print(f"🚀 Launching Optimizer (Mac): {safe_cmd}")
-                        # Open a NEW Terminal window (not just a tab) and bring it to front
-                        applescript = (
-                            'tell application "Terminal"\n'
-                            '    activate\n'
-                            f'    do script "{safe_cmd}"\n'
-                            'end tell'
-                        )
-                        subprocess.run(["osascript", "-e", applescript])
-                        
-                    self.send_response(204)
-                    self.end_headers()
-                except SecurityError as e:
-                    print(f"🚨 Security violation: {e}")
-                    self.send_error(403, "Forbidden")
-                except ValueError as e:
-                    print(f"❌ Validation error: {e}")
-                    self.send_error(400, str(e))
-                except Exception as e:
-                    print(f"❌ Error in compress endpoint: {e}")
-                    self.send_error(500, str(e))
-
-            elif self.path.startswith("/api/keep_optimized?"):
-                user_name = self.get_current_user()
-                if not user_name:
-                    self.send_error(401, "Unauthorized")
-                    return
-
-                try:
-                    params = parse_qs(urlparse(self.path).query)
-                    original_path = params.get("original", [None])[0]
-                    optimized_path = params.get("optimized", [None])[0]
-
-                    print(f"🔄 keep_optimized: original={original_path}")
-                    print(f"🔄 keep_optimized: optimized={optimized_path}")
-
-                    if original_path and optimized_path:
-                        orig_abs = os.path.abspath(original_path)
-                        opt_abs = os.path.abspath(optimized_path)
-
-                        print(f"🔄 keep_optimized: orig_abs={orig_abs} exists={os.path.exists(orig_abs)}")
-                        print(f"🔄 keep_optimized: opt_abs={opt_abs} exists={os.path.exists(opt_abs)}")
-                        
-                        if os.path.exists(opt_abs):
-                            # Move opt to original (replace)
-                            # We might want to handle extension changes if they differ, 
-                            # but "overwrite" usually implies keeping the original filename?
-                            # Actually, if we convert mkv -> mp4, "Keep" should probably keep the mp4 extension.
-                            # But for now, let's assume we replace the content of the original if same ext, 
-                            # or replace the file entirely if diff ext.
-                            
-                            # Safest: Remove original, move optimized to original name?
-                            # Wait, if we change ext, we should probably rename to original_stem.new_ext
-                            # But then we have a dangling original entry in cache.
-                            # Let's simple Rename optimized -> original_path (this forces original extension if we are not careful)
-                            
-                            # Better approach for mixed extensions:
-                            # 1. Delete original file.
-                            # 2. Rename optimized file to original_stem + optimized_ext.
-                            # 3. BUT user expects "Replace". If I have movie.mkv and movie_opt.mp4.
-                            #    If I keep opt, I expect movie.mp4? Or movie.mkv (container swap)?
-                            #    FFmpeg optimization usually keeps container or standardizes to mp4.
-                            #    Let's go with: Rename optimized to (original_dir / original_stem . optimized_ext).
-                            
-                            # However, to keep it simple and robust matching the user's "overwrite" mental model:
-                            # If we just replace the original file, we preserve the original entry in the DB/Cache key?
-                            # No, cache key is path.
-                            
-                            # Let's strictly follow: REPLACE original with optimized.
-                            # If extensions differ, we delete original, and rename optimized to original's stem + opt's extension.
-                            
-                            orig_path_obj = Path(orig_abs)
-                            opt_path_obj = Path(opt_abs)
-                            
-                            new_path = orig_path_obj.with_suffix(opt_path_obj.suffix)
-                            
-                            # Delete original
-                            if os.path.exists(orig_abs):
-                                os.remove(orig_abs)
-                            
-                            # Rename optimized to new path
-                            os.rename(opt_abs, new_path)
-                            
-                            # Update Cache
-                            
-                            # Remove optimized entry
-                            db.remove(opt_abs)
-                                
-                            # Remove old original entry if path changed (ext changed)
-                            if orig_abs != str(new_path):
-                                db.remove(orig_abs)
-                                
-                            db.save()
-                            _media_cache.invalidate()  # Einträge entfernt → Cache leeren
-                            
-                        else:
-                            print(f"❌ Optimized file not found: {opt_abs}")
-                        
-                    self.send_response(204)
-                    self.end_headers()
-                except Exception as e:
-                    print(f"❌ Error in keep_optimized: {e}")
-                    self.send_error(500, str(e))
-
-            elif self.path.startswith("/api/discard_optimized?"):
-                user_name = self.get_current_user()
-                if not user_name:
-                    self.send_error(401, "Unauthorized")
-                    return
-
-                try:
-                    params = parse_qs(urlparse(self.path).query)
-                    path = params.get("path", [None])[0]
-                    
-                    if path:
-                        abs_path = os.path.abspath(path)
-                        if os.path.exists(abs_path):
-                            os.remove(abs_path)
-                            db.remove(abs_path)
-                            db.save()
-                                
-                            # Regenerate Report
-                            try:
-                                current_port = self.server.server_address[1]
-                                report_debouncer.schedule(current_port)
-                            except Exception as e:
-                                print(f"⚠️ Report gen scheduling failed: {e}")
-                                
-                            print(f"🗑️ Discarded optimized: {os.path.basename(abs_path)}")
-                            
-                    self.send_response(204)
-                    self.end_headers()
-                except Exception as e:
-                    print(f"❌ Error in discard_optimized: {e}")
-                    self.send_error(500, str(e))
-
-            # --- RESCAN ---
-            elif self.path == "/api/rescan":
-                user_name = self.get_current_user()
-                if not user_name:
-                    self.send_error(401, "Unauthorized")
-                    return
-
-                print("🔄 Scan requested via API...")
-                try:
-                    # asyncio.run() creates a new event loop; safe to call from this
-                    # worker thread since SimpleHTTPServer runs each request in its own thread.
-                    import asyncio
-                    mgr = get_scanner_manager()
-                    loop = asyncio.new_event_loop()
-                    asyncio.set_event_loop(loop)
-                    try:
-                        new_count = loop.run_until_complete(mgr.run_scan())
-                    finally:
-                        loop.close()
-                        asyncio.set_event_loop(None)
-
-                    # Generate Report
-                    port = self.server.server_address[1]
-                    results = [e.model_dump(by_alias=True) for e in _media_cache.get()]
-                    _media_cache.invalidate()  # Neue Scan-Ergebnisse sofort sichtbar
-                    generate_html_report(results, config.report_file, server_port=port)
-
-                    self.send_response(200)
-                    self.send_header("Content-type", "application/json")
-                    self.end_headers()
-                    self.wfile.write(json.dumps({"status": "complete", "count": new_count}).encode())
-                    print("✅ Rescan complete.")
-
-                except Exception as e:
-                    print(f"❌ Rescan failed: {e}")
-                    self.send_error(500, str(e))
-
-            elif self.path == "/api/backup":
-                try:
-                    # Security Fix: Ensure only authenticated/local users (implicitly local)
-                    user_name = self.get_current_user()
-                    if not user_name:
-                        self.send_error(401, "Unauthorized")
-                        return
-
-                    print("💾 Backup requested...")
-                    
-                    # Force save first to ensure latest memory state is on disk?
-                    # config.save({}) # No-op save to flush? No, config.save updates logic.
-                    # Just read the file.
-                    
-                    if os.path.exists(SETTINGS_FILE):
-                        with open(SETTINGS_FILE, 'rb') as f:
-                            data = f.read()
-                            
-                        self.send_response(200)
-                        self.send_header("Content-Type", "application/json")
-                        self.send_header("Content-Disposition", 'attachment; filename="arcade_settings_backup.json"')
-                        self.end_headers()
-                        self.wfile.write(data)
-                        print("✅ Backup sent.")
-                    else:
-                        self.send_error(404, "Settings file not found")
-                        
-                except Exception as e:
-                    print(f"❌ Backup failed: {e}")
-                    self.send_error(500, str(e))
-
-            elif self.path.startswith("/batch_compress?paths="):
-                user_name = self.get_current_user()
-                if not user_name:
-                    self.send_error(401, "Unauthorized")
-                    return
-
-                try:
-                    # Use ||| as separator to avoid issues with commas in filenames
-                    paths = unquote(self.path.split("paths=")[1]).split("|||")
-                    current_port = self.server.server_address[1]
-                    
-                    # Validate all paths first
-                    validated_paths = []
-                    for p in paths:
-                        try:
-                            validated_path = sanitize_path(p)
-                            if os.path.exists(validated_path):
-                                validated_paths.append(validated_path)
-                            else:
-                                print(f"⚠️ Skipping non-existent file: {validated_path}")
-                        except (SecurityError, ValueError) as e:
-                            print(f"🚨 Skipping invalid path in batch: {p} - {e}")
-                            continue
-                    
-                    if not validated_paths:
-                        print("❌ No valid files to process in batch")
-                        self.send_response(204)
-                        self.end_headers()
-                        return
-                    
-                    # Build batch controller command
-                    batch_controller_path = os.path.join(
-                        os.path.dirname(config.optimizer_path), 
-                        "batch_controller.py"
-                    )
-                    
-                    # Comma-separate paths for the controller
-                    files_arg = ",".join(validated_paths)
-                    
-                    cmd_parts = [
-                        sys.executable,
-                        batch_controller_path,
-                        f"--files={files_arg}",
-                        f"--port={current_port}"
-                    ]
-                    
-                    if IS_WIN:
-                        # Windows: Launch in new console
-                        subprocess.Popen(cmd_parts, creationflags=subprocess.CREATE_NEW_CONSOLE)
-                    else:
-                        # macOS: Single terminal window
-                        safe_cmd = ' '.join(shlex.quote(str(p)) for p in cmd_parts)
-                        print(f"🚀 Launching Batch Controller: {len(validated_paths)} files")
-                        # Escape backslashes and quotes for AppleScript string
-                        escaped_cmd = safe_cmd.replace('\\', '\\\\').replace('"', '\\"')
-                        applescript = f'tell application "Terminal" to do script "{escaped_cmd}"'
-                        subprocess.run(["osascript", "-e", applescript])
-                    
-                    self.send_response(204)
-                    self.end_headers()
-                except Exception as e:
-                    print(f"❌ Error in batch_compress: {e}")
-                    self.send_error(500)
-            elif self.path.startswith("/hide?"):
-                user_name = self.get_current_user()
-                if not user_name:
-                    self.send_error(401)
-                    return
-                
-                params = parse_qs(urlparse(self.path).query)
-                path = params.get("path", [None])[0]
-                state = params.get("state", ["true"])[0].lower() == "true"
-                
-                if path:
-                    abs_path = os.path.abspath(path)
-                    u = user_db.get_user(user_name)
-                    if u:
-                        if state:
-                            if abs_path not in u.data.vaulted:
-                                u.data.vaulted.append(abs_path)
-                        else:
-                            if abs_path in u.data.vaulted:
-                                u.data.vaulted.remove(abs_path)
-                        user_db.add_user(u)
-                        print(f"Updated vault state for {user_name}: {os.path.basename(abs_path)} -> hidden={state}")
-                    
-                self.send_response(204)
-                self.end_headers()
-            elif self.path.startswith("/batch_hide?paths="):
-                user_name = self.get_current_user()
-                if not user_name:
-                    self.send_error(401)
-                    return
-                    
-                paths_str = unquote(self.path.split("paths=")[1])
-                paths_list = paths_str.split("&state=")[0].split(",")
-                state = "state=false" not in self.path
-                
-                u = user_db.get_user(user_name)
-                if u:
-                    updated_count = 0
-                    for p in paths_list:
-                        abs_p = os.path.abspath(p)
-                        if state:
-                            if abs_p not in u.data.vaulted:
-                                u.data.vaulted.append(abs_p)
-                                updated_count += 1
-                        else:
-                            if abs_p in u.data.vaulted:
-                                u.data.vaulted.remove(abs_p)
-                                updated_count += 1
-                    user_db.add_user(u)
-                    print(f"Batch updated vault state for {user_name} ({updated_count} files) -> hidden={state}")
-                self.send_response(204)
-                self.end_headers()
-            elif self.path.startswith("/favorite?"):
-                user_name = self.get_current_user()
-                if not user_name:
-                    self.send_error(401)
-                    return
-                    
-                params = parse_qs(urlparse(self.path).query)
-                path = params.get("path", [None])[0]
-                state = params.get("state", ["true"])[0].lower() == "true"
-                if path:
-                    abs_path = os.path.abspath(path)
-                    u = user_db.get_user(user_name)
-                    if u:
-                        if state:
-                            if abs_path not in u.data.favorites:
-                                u.data.favorites.append(abs_path)
-                        else:
-                            if abs_path in u.data.favorites:
-                                u.data.favorites.remove(abs_path)
-                        user_db.add_user(u)
-                        print(f"Updated favorite state for {user_name}: {os.path.basename(abs_path)} -> favorite={state}")
-                        
-                self.send_response(204)
-                self.end_headers()
-            elif self.path.startswith("/batch_favorite?"):
-                user_name = self.get_current_user()
-                if not user_name:
-                    self.send_error(401)
-                    return
-                
-                params = parse_qs(urlparse(self.path).query)
-                paths = params.get("paths", [""])[0].split(",")
-                state = params.get("state", ["true"])[0].lower() == "true"
-                
-                u = user_db.get_user(user_name)
-                if u:
-                    updated_count = 0
-                    for p in paths:
-                        if p:
-                            abs_path = os.path.abspath(p)
-                            if state:
-                                if abs_path not in u.data.favorites:
-                                    u.data.favorites.append(abs_path)
-                                    updated_count += 1
-                            else:
-                                if abs_path in u.data.favorites:
-                                    u.data.favorites.remove(abs_path)
-                                    updated_count += 1
-                    user_db.add_user(u)
-                    print(f"Batch updated favorite state for {user_name} ({updated_count} files) -> favorite={state}")
-                self.send_response(204)
-                self.end_headers()
+            # Routes below are handled by routes/files.py (registered above in the
+            # module dispatcher). The following elif chain is legacy code kept for
+            # reference during the transition phase. It is unreachable because
+            # files.handle_get() returns True for all these paths.
+            #
+            # Paths migrated to routes/files.py:
+            #   /reveal?         /api/mark_optimized?   /compress?
+            #   /api/keep_optimized?  /api/discard_optimized?
+            #   /api/rescan      /api/backup
+            #   /batch_compress?  /hide?  /batch_hide?  /favorite?  /batch_favorite?
             elif self.path.startswith("/stream?"):
                 try:
                     params = parse_qs(urlparse(self.path).query)
