+++ arcade_scanner/server/static/store.js
@@ -136,6 +136,7 @@ window.safeMode = localStorage.getItem('safe_mode') === 'true';
 window.renderedCount = 0;
 window.folderBrowserState = { currentPath: null, showVideosHere: false };
 window.folderBrowserPath = null;
+window.currentFolder = 'all';
 window.filteredVideos = [];
 
 console.log("🚀 Arcade Scanner State Manager Initialized");
